[level 2] Title: 멀쩡한 사각형, Time: 0.09 ms, Memory: 78.7 MB -BaekjoonHub

+++ 프로그래머스/2/62048. 멀쩡한 사각형/멀쩡한 사각형.java
@@ -1,13 +1,13 @@
 class Solution {
-    public long solution(long w, long h) {
-        long g = gcd(w, h);
-        long unusable = w + h - g;
-        return w * h - unusable;
+    public long solution(int w, int h) {
+        int g = gcd(w, h);
+        int unusable = w + h - g;
+        return (long) w * h - unusable;
     }
     
-    long gcd(long a, long b) {
+    int gcd(int a, int b) {
         while(b != 0) {
-            long temp = a % b;
+            int temp = a % b;
             a = b;
             b = temp;
         }
